// frontend/src/service/models/customer/customerUpdateReq.ts
export interface ICustomerItemUpdateReq {
  name: string;
  address: string;
  district: string;
  postcode: string;
  phone: string;
  email: string;
}

// 	Name     string `json:"name"`
// Address  string `json:"address"`
// District string `json:"district"`
// Postcode string `json:"postcode"`
// Phone    string `json:"phone"`
// Email    string `json:"email"`